Заменил файлы от Кекса на TS.
